Add `re-pattern` to builtins
Clojure can resolve functions from clojure/string so that you can
check for substrings

`[(clojure.string/includes? ?term "X")]`

Clojurescript lacks this runtime resolution so there is no way to
check for string containment. There are some helpful regex functions
but this requires the regex pattern to be constructed outside of the
query. This change allows to create the pattern in the query as
demonstrated in the test:

``` clojure
(d/q '[:find  ?name
       :in    [?name ...] ?key
       :where [(re-pattern ?key) ?pattern]
              [(re-find ?pattern ?name)]]
     #{"abc" "abcX" "aXb"}
     "X")
```