Query: shortcircuit clause resolution when result is guaranteed to be empty.
The `-collect` fn, run at the end of a query, correctly shortcircuits with the
following comment: "one empty rel means final set has to be empty"

However, while processing the query clauses, there was no shortcircuit.
Also, since the implementation of `lookup-pattern` is agnostic to
existing rels, this means that attempts to write short-circuiting
clauses at the top of a complex query have very little effect: the
query engine will continue querying the indexes for every tuple that
could possibly match every clause.

This adds a shortcircuit in `resolve-clause` if any existing relation
is empty. The check is cheap, and it should provide a substantial
performance boost in many common cases.